+++ lib/ResponsiveReactGridLayout.jsx
@@ -142,7 +142,7 @@ export default class ResponsiveReactGridLayout extends React.Component<
     };
   }
 
-  componentWillReceiveProps(nextProps: Props<*>) {
+  UNSAFE_componentWillReceiveProps(nextProps: Props<*>) {
     // Allow parent to set width or breakpoint directly.
     if (
       nextProps.width != this.props.width ||